+++ src/ralph/models/xapi/navigation/fields/objects.py
@@ -35,4 +35,4 @@ class PageObjectField(ObjectField):
         definition (PageObjectDefinitionField): See PageObjectDefinitionField.
     """
 
-    definition: PageObjectDefinitionField
+    definition: PageObjectDefinitionField = PageObjectDefinitionField()