+++ tests/test_utils.py
@@ -1,6 +1,5 @@
 """Tests for Ralph utils"""
 
-
 import pytest
 
 from ralph import utils as ralph_utils
@@ -98,7 +97,7 @@ def test_utils_get_instance_from_class():
 
 @pytest.mark.parametrize("path,value", [(["foo", "bar"], "bar_value")])
 def test_utils_get_dict_value_from_path_should_return_given_value(path, value):
-    """The get_dict_value_from_path function should return the value when it's present."""
+    """Tests the get_dict_value_from_path function should return the value when it's present."""
 
     dictionary = {"foo": {"bar": "bar_value"}}
     assert ralph_utils.get_dict_value_from_path(dictionary, path) == value
@@ -116,7 +115,23 @@ def test_utils_get_dict_value_from_path_should_return_given_value(path, value):
 def test_utils_get_dict_value_from_path_should_return_none_when_value_does_not_exists(
     path,
 ):
-    """The get_dict_value_from_path function should return None if the value is not found."""
+    """Tests the get_dict_value_from_path function should return None if the value is not found."""
 
     dictionary = {"foo": {"bar": "bar_value"}}
     assert ralph_utils.get_dict_value_from_path(dictionary, path) is None
+
+
+def test_utils_set_dict_value_from_path_creating_new_fields():
+    """Tests when the fields are not present, set_dict_value_from_path should add them."""
+
+    dictionary = {}
+    ralph_utils.set_dict_value_from_path(dictionary, ["foo", "bar"], "baz")
+    assert dictionary == {"foo": {"bar": "baz"}}
+
+
+def test_utils_set_dict_value_from_path_updating_fields():
+    """Tests when the fields are present, set_dict_value_from_path should update them."""
+
+    dictionary = {"foo": {"bar": "bar_value"}}
+    ralph_utils.set_dict_value_from_path(dictionary, ["foo", "bar"], "baz")
+    assert dictionary == {"foo": {"bar": "baz"}}